requirements.txt
livereload==2.6.3
Flask==2.0.1
mysql_connector_repackaged==0.3.1
